+++ packages/solecs/src/MapSet.sol
@@ -6,8 +6,8 @@ pragma solidity >=0.8.0;
  */
 contract MapSet {
   address private owner;
-  mapping(uint256 => uint256[]) private items;
-  mapping(uint256 => mapping(uint256 => uint256)) private itemToIndex;
+  mapping(uint256 => uint256[]) internal items;
+  mapping(uint256 => mapping(uint256 => uint256)) internal itemToIndex;
 
   constructor() {
     owner = msg.sender;